solution/pyproject.toml
[tool.poetry]
name = "Pulse Backend"
version = "0.1.0"
description = ""
authors = ["Погонцев Данил Сергеевич"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.10"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
psycopg2 = "^2.9.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "ARG002", "A002", "RUF001"]

[tool.ruff.lint.per-file-ignores]
"*/models.py" = ["N815"]
"*/serializers.py" = ["N815"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
